requirements.txt
greenlet==3.1.1
msgpack==1.1.0
packaging==24.2
pyserial==3.5
python-can==4.5.0
redis==5.2.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
wrapt==1.17.2
